Move redirects to process file
