+++ packages/cli/src/process.ts
@@ -79,3 +79,19 @@ export function killAll(signal: NodeJS.Signals) {
   }
   childProcesses = []
 }
+
+export function redirectToServer(argv: string[]) {
+  spawn("npm", ["exec", "--", "phero-server", ...argv], {
+    cwd: process.cwd(),
+    detached: false,
+    stdio: "inherit",
+  })
+}
+
+export function redirectToClient(argv: string[]) {
+  spawn("npm", ["exec", "--", "phero-client", ...argv], {
+    cwd: process.cwd(),
+    detached: false,
+    stdio: "inherit",
+  })
+}
